Split best-for detail helpers and view
